requirements.txt
streamlit==1.32.0
streamlit-webrtc==0.47.1
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
pandas==2.2.1
av==16.1.0
plotly==5.20.0
anthropic==0.28.0
groq==0.5.0
python-dotenv==1.0.1
bcrypt==4.1.2
requests==2.31.0
